+++ src/TSVNCache/DirectoryWatcher.h
@@ -1,6 +1,6 @@
 // TortoiseSVN - a Windows shell extension for easy version control
 
-// External Cache Copyright (C) 2005-2008, 2012 - TortoiseSVN
+// External Cache Copyright (C) 2005-2008, 2012, 2014 - TortoiseSVN
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -113,7 +113,7 @@ private:
     CTSVNPathList           watchedPaths;   ///< list of watched paths.
 
     CTSVNPath               blockedPath;
-    DWORD                   blockTickCount;
+    ULONGLONG               blockTickCount;
 
     /**
      * \ingroup TSVNCache
