+++ src/TSVNCache/SVNStatusCache.h
@@ -1,6 +1,6 @@
 // TortoiseSVN - a Windows shell extension for easy version control
 
-// Copyright (C) 2005-2006, 2013 - TortoiseSVN
+// Copyright (C) 2005-2006, 2013-2014 - TortoiseSVN
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -94,7 +94,7 @@ public:
     bool IsPathGood(const CTSVNPath& path);
     bool IsPathWatched(const CTSVNPath& path) {return watcher.IsPathWatched(path);}
     bool AddPathToWatch(const CTSVNPath& path) {return watcher.AddPath(path);}
-    bool BlockPath(const CTSVNPath& path, bool specific, DWORD timeout = 0);
+    bool BlockPath(const CTSVNPath& path, bool specific, ULONGLONG timeout = 0);
     bool UnBlockPath(const CTSVNPath& path);
     bool RemoveTimedoutBlocks();
 
@@ -109,7 +109,7 @@ private:
     CCachedDirectory::CachedDirMap m_directoryCache;
 
     CComAutoCriticalSection m_NoWatchPathCritSec;
-    std::map<CTSVNPath, DWORD> m_NoWatchPaths;  ///< paths to block from getting crawled, and the time in ms until they're unblocked
+    std::map<CTSVNPath, ULONGLONG> m_NoWatchPaths;  ///< paths to block from getting crawled, and the time in ms until they're unblocked
 
     SVNHelper m_svnHelp;
     ShellCache  m_shellCache;
@@ -123,7 +123,7 @@ private:
     CComAutoCriticalSection m_critSec;
     CTSVNPath m_mostRecentAskedPath;
     CStatusCacheEntry m_mostRecentStatus;
-    long m_mostRecentExpiresAt;
+    LONGLONG m_mostRecentExpiresAt;
 
     CDirectoryWatcher watcher;
