+++ src/TortoiseShell/ShellExt.h
@@ -1,6 +1,6 @@
 // TortoiseSVN - a Windows shell extension for easy version control
 
-// Copyright (C) 2003-2013 - TortoiseSVN
+// Copyright (C) 2003-2014 - TortoiseSVN
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -31,7 +31,7 @@ extern  volatile LONG       g_cRefThisDll;          // Reference count of this D
 extern  HINSTANCE           g_hmodThisDll;          // Instance handle for this DLL
 extern  ShellCache          g_ShellCache;           // caching of registry entries, ...
 extern  DWORD               g_langid;
-extern  DWORD               g_langTimeout;
+extern  ULONGLONG           g_langTimeout;
 extern  HINSTANCE           g_hResInst;
 extern  tstring             g_filepath;
 extern  svn_wc_status_kind  g_filestatus;           ///< holds the corresponding status to the file/dir above
