+++ src/Utils/Hooks.h
@@ -1,6 +1,6 @@
 // TortoiseSVN - a Windows shell extension for easy version control
 
-// Copyright (C) 2006-2013 - TortoiseSVN
+// Copyright (C) 2006-2014 - TortoiseSVN
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -318,7 +318,7 @@ private:
     bool                ApproveHook(HWND hWnd, hookiterator it);
 
     static CHooks *     m_pInstance;
-    DWORD               m_lastPreConnectTicks;
+    ULONGLONG           m_lastPreConnectTicks;
     bool                m_PathsConvertedToUrls;
     CTSVNPath           m_wcRootPath;
 };
